travel_assistant_code_git/requirements.txt
# Core Django and Web
Django==5.2
asgiref==3.8.1
channels==4.1.0
django-cors-headers==4.4.0
django-environ==0.11.2
gunicorn==23.0.0

# AI/LLM Frameworks
langchain==0.3.26
langchain_openai==0.3.28
langchain_core==0.3.68
langchain_community==0.3.27
langchain_postgres==0.0.15
langchain_text_splitters==0.3.8
openai==1.93.0
tavily_python==0.7.9

# Data Handling & Processing
pandas==2.3.1
python-dotenv==1.1.1
openpyxl==3.0.10
xlsxwriter
python-dateutil==2.9.0.post0

# Document Processing (PDFs)
PyMuPDF==1.25.4
pdfminer.six==20231228
pdfplumber==0.11.5

# Utility & Helpers
python-decouple==3.8
python-slugify
fuzzywuzzy
